# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CollectionCreateRequest": {
        "properties": {
          "department": {
            "default": "",
            "title": "Department",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "email": {
            "default": "",
            "title": "Email",
            "type": "string"
          },
          "folder": {
            "default": "",
            "title": "Folder",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "password": {
            "default": "",
            "title": "Password",
            "type": "string"
          },
          "read_teams": {
            "items": {
              "type": "string"
            },
            "title": "Read Teams",
            "type": "array"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "subfolder": {
            "default": "",
            "title": "Subfolder",
            "type": "string"
          }
        },
        "title": "CollectionCreateRequest",
        "type": "object"
      },
      "CollectionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CollectionResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "title": "CollectionListResponse",
        "type": "object"
      },
...
      "CollectionResponse": {
        "properties": {
          "collection_id": {
            "title": "Collection Id",
            "type": "string"
          },
          "department": {
            "title": "Department",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "folder": {
            "default": "",
            "title": "Folder",
            "type": "string"
          },
          "job_ids": {
            "items": {
              "type": "string"
            },
            "title": "Job Ids",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "read_teams": {
            "items": {
              "type": "string"
            },
            "title": "Read Teams",
            "type": "array"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "subfolder": {
            "default": "",
            "title": "Subfolder",
            "type": "string"
          }
        },
        "required": [
          "collection_id",
          "slug",
          "name",
          "email",
          "department"
        ],
        "title": "CollectionResponse",
        "type": "object"
      },
...
      "CollectionUpdateRequest": {
        "description": "PATCH payload: every field optional and only applied when present\n(None = leave unchanged). `slug` is deliberately not patchable here --\nit is the collection's stable cross-service identity (see Collection's\ndocstring) and already-processed documents carry it in their frontmatter.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "read_teams": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read Teams"
          }
        },
        "title": "CollectionUpdateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Weave Ingest API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/collections": {
      "get": {
        "description": "Same visibility rule as GET /jobs (`_visible_job_filter`): own +\ncurrent-teammates' + (for an admin) every collection. Does not populate\n`job_ids` per item -- see CollectionResponse's docstring -- use GET\n/collections/{id} for a single collection's job membership.",
        "operationId": "list_collections_api_v1_collections_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Collections"
      },
      "post": {
        "operationId": "create_collection_api_v1_collections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Collection"
      }
    },
...
    "/api/v1/collections/{collection_id}": {
      "get": {
        "operationId": "get_collection_api_v1_collections__collection_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "title": "Collection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Collection"
      },
      "patch": {
        "operationId": "update_collection_api_v1_collections__collection_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "title": "Collection Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Collection"
      }
    },
...
  }
}